patches: bypass_checksum: hook up confirm()

+++ MBCPApp/patches/bypass_checksum.sh
@@ -8,6 +8,8 @@
 mkdir tools/bypass
 warn "Source code for [app-debug.apk] are on [https://git.disroot.org/cuynu/ApkSignatureKillerEx]"
 warn "Ensure that you applied all patches you want, and then apply this as the last one!"
+confirm "Are you sure you want to apply this patch?" || exit 69
+
 info "Downloading pre-compiled ApkSignatureKillerEx..."
 wget -q --show-progress -O 'tools/bypass/app-debug.apk' 'https://git.disroot.org/cuynu/ApkSignatureKillerEx/releases/download/1.0/app-debug.apk'
 info "Extracting [app-debug.apk]..."
